requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.3
billiard==3.3.0.23
celery==3.1.18
dj-database-url==0.5.0
Django==3.0.3
django-background-tasks==1.2.5
django-celery==3.3.1
django-common-helpers==0.9.2
django-compat==1.0.15
django-crispy-forms==1.8.1
django-cron==0.5.1
django-environ==0.4.5
django-heroku==0.3.1
django-multiselectfield==0.1.12
gunicorn==20.0.4
kombu==3.0.37
Pillow==7.1.2
psycopg2==2.8.4
pytz==2019.3
redis==3.5.3
six==1.15.0
sqlparse==0.3.0
vine==1.3.0
whitenoise==5.0.1
